// server/data-source/slack/slack-reply.interface.ts
export interface SlackMessage {
  client_msg_id?: string;
  type: string;
  text: string;
  user: string;
  ts: string;
  blocks: Array<{
    type: string;
    block_id: string;
    elements: Array<{
      type: string;
      elements: Array<{
        type: string;
        text: string;
      }>;
    }>;
  }>;
  team?: string;
  thread_ts?: string;
  parent_user_id?: string;
  reactions?: Array<{
    name: string;
    users: string[];
    count: number;
  }>;
  files?: Array<{
    id: string;
    created: number;
    timestamp: number;
    name: string;
    title: string;
    mimetype: string;
    filetype: string;
    pretty_type: string;
    user: string;
    user_team: string;
    editable: boolean;
    size: number;
    mode: string;
    is_external: boolean;
    external_type: string;
    is_public: boolean;
    public_url_shared: boolean;
    display_as_bot: boolean;
    username: string;
    url_private: string;
    url_private_download: string;
    media_display_type: string;
    thumb_64: string;
    thumb_80: string;
    thumb_360: string;
    thumb_360_w: number;
    thumb_360_h: number;
    thumb_480: string;
    thumb_480_w: number;
    thumb_480_h: number;
    thumb_160: string;
    thumb_720: string;
    thumb_720_w: number;
    thumb_720_h: number;
    thumb_800: string;
    thumb_800_w: number;
    thumb_800_h: number;
    thumb_960: string;
    thumb_960_w: number;
    thumb_960_h: number;
    thumb_1024: string;
    thumb_1024_w: number;
    thumb_1024_h: number;
    original_w: number;
    original_h: number;
    thumb_tiny: string;
    permalink: string;
    permalink_public: string;
    is_starred: boolean;
    has_rich_preview: boolean;
    file_access: string;
  }>;
  upload?: boolean;
  reply_count?: number;
  reply_users_count?: number;
  latest_reply?: string;
  reply_users?: string[];
  is_locked?: boolean;
  subscribed?: boolean;
}